requirements.txt
azure-ai-projects==1.0.0b10
semantic-kernel[azure]
azure-identity
fastapi
uvicorn
python-dotenv
gunicorn